Common utilities moved to FingerprintUtil.h and code duplication reduced
