+++ Code/GraphMol/Fingerprints/CMakeLists.txt
@@ -4,7 +4,7 @@ add_definitions(-DRDKIT_FINGERPRINTS_BUILD)
 rdkit_library(Fingerprints
               Fingerprints.cpp PatternFingerprints.cpp MorganFingerprints.cpp
               AtomPairs.cpp MACCS.cpp AtomPairGenerator.cpp
-              MorganGenerator.cpp RDKitFPGenerator.cpp
+              MorganGenerator.cpp RDKitFPGenerator.cpp FingerprintUtil.cpp
               LINK_LIBRARIES Subgraphs SubstructMatch SmilesParse GraphMol
               ${RDKit_THREAD_LIBS} )
 
@@ -16,6 +16,7 @@ rdkit_headers(AtomPairs.h
               AtomPairGenerator.h
               MorganGenerator.h
               RDKitFPGenerator.h
+              FingerprintUtil.h
               DEST GraphMol/Fingerprints)
 
 rdkit_test(testFingerprints test1.cpp LINK_LIBRARIES
