+++ Code/GraphMol/Fingerprints/Fingerprints.cpp
@@ -27,6 +27,7 @@
 #include <RDGeneral/types.h>
 #include <algorithm>
 #include <boost/dynamic_bitset.hpp>
+#include <GraphMol/Fingerprints/FingerprintUtil.h>
 
 //#define VERBOSE_FINGERPRINTING 1
 //#define REPORT_FP_STATS 1
@@ -172,173 +173,6 @@ return res;
 
 }  // end of anonymous namespace
 
-namespace utils {
-
-void buildDefaultRDKitFingerprintAtomInvariants(
-    const ROMol &mol, std::vector<boost::uint32_t> &lAtomInvariants) {
-  lAtomInvariants.clear();
-  lAtomInvariants.reserve(mol.getNumAtoms());
-  for (ROMol::ConstAtomIterator atomIt = mol.beginAtoms();
-       atomIt != mol.endAtoms(); ++atomIt) {
-    unsigned int aHash = ((*atomIt)->getAtomicNum() % 128) << 1 |
-                         static_cast<unsigned int>((*atomIt)->getIsAromatic());
-    lAtomInvariants.push_back(aHash);
-  }
-}
-
-void enumerateAllPaths(const ROMol &mol, INT_PATH_LIST_MAP &allPaths,
-                       const std::vector<boost::uint32_t> *fromAtoms,
-                       bool branchedPaths, bool useHs, unsigned int minPath,
-                       unsigned int maxPath) {
-  if (!fromAtoms) {
-    if (branchedPaths) {
-      allPaths = findAllSubgraphsOfLengthsMtoN(mol, minPath, maxPath, useHs);
-    } else {
-      allPaths = findAllPathsOfLengthsMtoN(mol, minPath, maxPath, true, useHs);
-    }
-  } else {
-    BOOST_FOREACH (boost::uint32_t aidx, *fromAtoms) {
-      INT_PATH_LIST_MAP tPaths;
-      if (branchedPaths) {
-        tPaths =
-            findAllSubgraphsOfLengthsMtoN(mol, minPath, maxPath, useHs, aidx);
-      } else {
-        tPaths =
-            findAllPathsOfLengthsMtoN(mol, minPath, maxPath, true, useHs, aidx);
-      }
-      for (INT_PATH_LIST_MAP::const_iterator tpit = tPaths.begin();
-           tpit != tPaths.end(); ++tpit) {
-#ifdef VERBOSE_FINGERPRINTING
-        std::cerr << "paths from " << aidx << " size: " << tpit->first
-                  << std::endl;
-        BOOST_FOREACH (PATH_TYPE path, tpit->second) {
-          std::cerr << " path: ";
-          std::copy(path.begin(), path.end(),
-                    std::ostream_iterator<int>(std::cerr, ", "));
-          std::cerr << std::endl;
-        }
-#endif
-
-        allPaths[tpit->first].insert(allPaths[tpit->first].begin(),
-                                     tpit->second.begin(), tpit->second.end());
-      }
-    }
-  }
-}
-
-void identifyQueryBonds(const ROMol &mol, std::vector<const Bond *> &bondCache,
-                        std::vector<short> &isQueryBond) {
-  bondCache.resize(mol.getNumBonds());
-  ROMol::EDGE_ITER firstB, lastB;
-  boost::tie(firstB, lastB) = mol.getEdges();
-  while (firstB != lastB) {
-    const Bond *bond = mol[*firstB];
-    isQueryBond[bond->getIdx()] = 0x0;
-    bondCache[bond->getIdx()] = bond;
-    if (isComplexQuery(bond)) {
-      isQueryBond[bond->getIdx()] = 0x1;
-    }
-    if (isComplexQuery(bond->getBeginAtom())) {
-      isQueryBond[bond->getIdx()] |= 0x2;
-    }
-    if (isComplexQuery(bond->getEndAtom())) {
-      isQueryBond[bond->getIdx()] |= 0x4;
-    }
-    ++firstB;
-  }
-}
-
-std::vector<unsigned int> generateBondHashes(
-    const ROMol &mol, boost::dynamic_bitset<> &atomsInPath,
-    const std::vector<const Bond *> &bondCache,
-    const std::vector<short> &isQueryBond, const PATH_TYPE &path,
-    bool useBondOrder, std::vector<boost::uint32_t> *atomInvariants) {
-  PRECONDITION(!atomInvariants || atomInvariants->size() >= mol.getNumAtoms(),
-               "bad atomInvariants size");
-
-  std::vector<unsigned int> bondHashes;
-  atomsInPath.reset();
-  bool queryInPath = false;
-  std::vector<unsigned int> atomDegrees(mol.getNumAtoms(), 0);
-  for (unsigned int i = 0; i < path.size() && !queryInPath; ++i) {
-    const Bond *bi = bondCache[path[i]];
-    CHECK_INVARIANT(bi, "bond not in cache");
-    atomDegrees[bi->getBeginAtomIdx()]++;
-    atomDegrees[bi->getEndAtomIdx()]++;
-    atomsInPath.set(bi->getBeginAtomIdx());
-    atomsInPath.set(bi->getEndAtomIdx());
-    if (isQueryBond[path[i]]) queryInPath = true;
-  }
-  if (queryInPath) {
-    return bondHashes;
-  }
-
-  // -----------------
-  // calculate the bond hashes:
-  std::vector<unsigned int> bondNbrs(path.size(), 0);
-  bondHashes.reserve(path.size() + 1);
-
-  for (unsigned int i = 0; i < path.size(); ++i) {
-    const Bond *bi = bondCache[path[i]];
-#ifdef REPORT_FP_STATS
-    if (std::find(atomsToUse.begin(), atomsToUse.end(),
-                  bi->getBeginAtomIdx()) == atomsToUse.end()) {
-      atomsToUse.push_back(bi->getBeginAtomIdx());
-    }
-    if (std::find(atomsToUse.begin(), atomsToUse.end(), bi->getEndAtomIdx()) ==
-        atomsToUse.end()) {
-      atomsToUse.push_back(bi->getEndAtomIdx());
-    }
-#endif
-    for (unsigned int j = i + 1; j < path.size(); ++j) {
-      const Bond *bj = bondCache[path[j]];
-      if (bi->getBeginAtomIdx() == bj->getBeginAtomIdx() ||
-          bi->getBeginAtomIdx() == bj->getEndAtomIdx() ||
-          bi->getEndAtomIdx() == bj->getBeginAtomIdx() ||
-          bi->getEndAtomIdx() == bj->getEndAtomIdx()) {
-        ++bondNbrs[i];
-        ++bondNbrs[j];
-      }
-    }
-#ifdef VERBOSE_FINGERPRINTING
-    std::cerr << "   bond(" << i << "):" << bondNbrs[i] << std::endl;
-#endif
-    // we have the count of neighbors for bond bi, compute its hash:
-    unsigned int a1Hash = (*atomInvariants)[bi->getBeginAtomIdx()];
-    unsigned int a2Hash = (*atomInvariants)[bi->getEndAtomIdx()];
-    unsigned int deg1 = atomDegrees[bi->getBeginAtomIdx()];
-    unsigned int deg2 = atomDegrees[bi->getEndAtomIdx()];
-    if (a1Hash < a2Hash) {
-      std::swap(a1Hash, a2Hash);
-      std::swap(deg1, deg2);
-    } else if (a1Hash == a2Hash && deg1 < deg2) {
-      std::swap(deg1, deg2);
-    }
-    unsigned int bondHash = 1;
-    if (useBondOrder) {
-      if (bi->getIsAromatic() || bi->getBondType() == Bond::AROMATIC) {
-        // makes sure aromatic bonds always hash as aromatic
-        bondHash = Bond::AROMATIC;
-      } else {
-        bondHash = bi->getBondType();
-      }
-    }
-    boost::uint32_t ourHash = bondNbrs[i];
-    gboost::hash_combine(ourHash, bondHash);
-    gboost::hash_combine(ourHash, a1Hash);
-    gboost::hash_combine(ourHash, deg1);
-    gboost::hash_combine(ourHash, a2Hash);
-    gboost::hash_combine(ourHash, deg2);
-    bondHashes.push_back(ourHash);
-    // std::cerr<<"    "<<bi->getIdx()<<"
-    // "<<a1Hash<<"("<<deg1<<")"<<"-"<<a2Hash<<"("<<deg2<<")"<<" "<<bondHash<<"
-    // -> "<<ourHash<<std::endl;
-  }
-  return bondHashes;
-}
-
-}  // end of namespace utils
-
 // caller owns the result, it must be deleted
 ExplicitBitVect *RDKFingerprintMol(
     const ROMol &mol, unsigned int minPath, unsigned int maxPath,
@@ -381,7 +215,7 @@ ExplicitBitVect *RDKFingerprintMol(
   // build default atom invariants if need be:
   std::vector<boost::uint32_t> lAtomInvariants;
   if (!atomInvariants) {
-    utils::buildDefaultRDKitFingerprintAtomInvariants(mol, lAtomInvariants);
+    RDKitFPUtils::buildDefaultRDKitFingerprintAtomInvariants(mol, lAtomInvariants);
     atomInvariants = &lAtomInvariants;
   }
 
@@ -389,13 +223,13 @@ ExplicitBitVect *RDKFingerprintMol(
 
   // get all paths
   INT_PATH_LIST_MAP allPaths;
-  utils::enumerateAllPaths(mol, allPaths, fromAtoms, branchedPaths, useHs,
+  RDKitFPUtils::enumerateAllPaths(mol, allPaths, fromAtoms, branchedPaths, useHs,
                            minPath, maxPath);
 
   // identify query bonds
   std::vector<short> isQueryBond(mol.getNumBonds(), 0);
   std::vector<const Bond *> bondCache;
-  utils::identifyQueryBonds(mol, bondCache, isQueryBond);
+  RDKitFPUtils::identifyQueryBonds(mol, bondCache, isQueryBond);
 
   if (atomBits) {
     for (unsigned int i = 0; i < mol.getNumAtoms(); ++i) {
@@ -430,7 +264,7 @@ ExplicitBitVect *RDKFingerprintMol(
 #if 1
       // the bond hashes of the path
       std::vector<unsigned int> bondHashes =
-          utils::generateBondHashes(mol, atomsInPath, bondCache, isQueryBond,
+          RDKitFPUtils::generateBondHashes(mol, atomsInPath, bondCache, isQueryBond,
                                     path, useBondOrder, atomInvariants);
       if (!bondHashes.size()) {
         continue;
@@ -885,19 +719,19 @@ SparseIntVect<boost::uint64_t> *getUnfoldedRDKFingerprintMol(
   // build default atom invariants if need be:
   std::vector<boost::uint32_t> lAtomInvariants;
   if (!atomInvariants) {
-    utils::buildDefaultRDKitFingerprintAtomInvariants(mol, lAtomInvariants);
+    RDKitFPUtils::buildDefaultRDKitFingerprintAtomInvariants(mol, lAtomInvariants);
     atomInvariants = &lAtomInvariants;
   }
 
   // get all paths
   INT_PATH_LIST_MAP allPaths;
-  utils::enumerateAllPaths(mol, allPaths, fromAtoms, branchedPaths, useHs,
+  RDKitFPUtils::enumerateAllPaths(mol, allPaths, fromAtoms, branchedPaths, useHs,
                            minPath, maxPath);
 
   // identify query bonds
   std::vector<short> isQueryBond(mol.getNumBonds(), 0);
   std::vector<const Bond *> bondCache;
-  utils::identifyQueryBonds(mol, bondCache, isQueryBond);
+  RDKitFPUtils::identifyQueryBonds(mol, bondCache, isQueryBond);
 
   if (atomBits) {
     for (unsigned int i = 0; i < mol.getNumAtoms(); ++i) {
@@ -913,7 +747,7 @@ SparseIntVect<boost::uint64_t> *getUnfoldedRDKFingerprintMol(
     BOOST_FOREACH (const PATH_TYPE &path, paths->second) {
       // the bond hashes of the path
       std::vector<unsigned int> bondHashes =
-          utils::generateBondHashes(mol, atomsInPath, bondCache, isQueryBond,
+          RDKitFPUtils::generateBondHashes(mol, atomsInPath, bondCache, isQueryBond,
                                     path, useBondOrder, atomInvariants);
       if (!bondHashes.size()) {
         continue;
